Revert "[Support] Move HTTP client/server to new LLVMSupportHTTP lib (NFC) (#…"
This reverts commit 04dbbd414f542d02cb0eaee79fe8a4f2e32d18bf.
