+++ llvm/include/llvm/Debuginfod/HTTPClient.h
@@ -1,4 +1,4 @@
-//===----------------------------------------------------------------------===//
+//===-- llvm/Support/HTTPClient.h - HTTP client library ---------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -12,8 +12,8 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_SUPPORT_HTTP_HTTPCLIENT_H
-#define LLVM_SUPPORT_HTTP_HTTPCLIENT_H
+#ifndef LLVM_DEBUGINFOD_HTTPCLIENT_H
+#define LLVM_DEBUGINFOD_HTTPCLIENT_H
 
 #include "llvm/ADT/SmallString.h"
 #include "llvm/ADT/SmallVector.h"
@@ -85,4 +85,4 @@ public:
 
 } // end namespace llvm
 
-#endif // LLVM_SUPPORT_HTTP_HTTPCLIENT_H
+#endif // LLVM_DEBUGINFOD_HTTPCLIENT_H
