+++ llvm/lib/Debuginfod/CMakeLists.txt
@@ -1,3 +1,13 @@
+# Link LibCURL if the user wants it
+if (LLVM_ENABLE_CURL)
+  set(imported_libs CURL::libcurl)
+endif()
+
+# Link cpp-httplib if the user wants it
+if (LLVM_ENABLE_HTTPLIB)
+  set(imported_libs ${imported_libs} httplib::httplib)
+endif()
+
 # Make sure pthread is linked if this is a unix host
 if (CMAKE_HOST_UNIX)
   set(imported_libs ${imported_libs} ${LLVM_PTHREAD_LIB})
@@ -8,6 +18,8 @@ endif()
 add_llvm_library(LLVMDebuginfod
   BuildIDFetcher.cpp
   Debuginfod.cpp
+  HTTPClient.cpp
+  HTTPServer.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/Debuginfod
@@ -17,7 +29,6 @@ add_llvm_library(LLVMDebuginfod
 
   LINK_COMPONENTS
   Support
-  SupportHTTP
   Symbolize
   DebugInfoDWARF
   BinaryFormat
