+++ llvm/lib/Debuginfod/HTTPClient.cpp
@@ -1,4 +1,4 @@
-//===--- HTTPClient.cpp - HTTP client library -----------------------------===//
+//===-- llvm/Debuginfod/HTTPClient.cpp - HTTP client library ----*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -12,13 +12,11 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "llvm/Support/HTTP/HTTPClient.h"
-
+#include "llvm/Debuginfod/HTTPClient.h"
 #include "llvm/ADT/APInt.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/Errc.h"
 #include "llvm/Support/Error.h"
-#include "llvm/Support/ManagedStatic.h"
 #include "llvm/Support/MemoryBuffer.h"
 #ifdef LLVM_ENABLE_CURL
 #include <curl/curl.h>
@@ -41,7 +39,7 @@ class HTTPClientCleanup {
 public:
   ~HTTPClientCleanup() { HTTPClient::cleanup(); }
 };
-ManagedStatic<HTTPClientCleanup> Cleanup;
+static const HTTPClientCleanup Cleanup;
 
 #ifdef LLVM_ENABLE_CURL
