+++ llvm/unittests/Debuginfod/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_llvm_unittest(DebuginfodTests
+  HTTPServerTests.cpp
   DebuginfodTests.cpp
   )
 
 target_link_libraries(DebuginfodTests PRIVATE
   LLVMDebuginfod
-  LLVMSupportHTTP
   LLVMTestingSupport
   )
