--- llvm/unittests/Support/HTTP/CMakeLists.txt
@@ -1,8 +1,0 @@
-add_llvm_unittest(SupportHTTPTests
-  HTTPServerTests.cpp
-  )
-
-target_link_libraries(SupportHTTPTests PRIVATE
-  LLVMSupportHTTP
-  LLVMTestingSupport
-  )
